refactor: move Triton-specific codegen into `helion/_compiler/triton/`
Relocate every Triton backend `@_decorators.codegen(op, "triton")` body -- and
the triton-only helpers/classes/constants they use -- out of the backend-neutral
language modules into per-module files under the Triton backend folder:

  `triton_atomic_ops.py`  (8 atomics + `_codegen_common`)
  `triton_matmul_ops.py`  (`dot`, `dot_scaled`)
  `triton_memory_ops.py`  (`store`, `load` + `_EVICTION_POLICY_MAP`)
  `triton_tracing_ops.py` (`_mask_to`)
  `triton_gelu_tanh_approx.py` (`_gelu_tanh_approx`, `_gelu_erf` + expr templates)
  `triton_view_ops.py`    (`split`, `join`)
  `triton_reduce_ops.py`  (`_reduce` + helper-fn emitters)
  `triton_scan_ops.py`    (`_associative_scan` + helper-fn emitters)
  `triton_quantized_ops.py` (`float4_e2m1fn_x2_to_float32` + FP4 asm)
  `triton_barrier.py` / `triton_debug_ops.py` / `triton_device_print.py` /
  `triton_inline_asm_ops.py` / `triton_inline_triton_ops.py` (`barrier`,
  `breakpoint`, `device_print`, `inline_asm_elementwise`, `inline_triton`,
  `triton_kernel` + the inline-triton source-manipulation helper suite)

Each language module keeps the backend-neutral op + a bottom
`from .._compiler.triton import triton_<module>  # noqa: E402, F401`
(after the pallas import where present) so codegens register with the same eager
timing. Shared helpers used by other backends (`_to_ast_values`,
`_maybe_materialize_tile_index_load`, `_needs_f32_accumulator`, `GELU_*`
constants, `_validate_args`) stay in the language module and are imported into the
triton file.

No behavior change: generated Triton code for a loop+dot+store+load kernel is
byte-identical to before; all 26 triton codegens now resolve from
`helion._compiler.triton.*`; triton execution tests pass (atomic/reduce/inline =
50 passed); pallas/cute/metal/common codegens untouched.

stack-info: PR: https://github.com/pytorch/helion/pull/3050, branch: AmesingFlank/stack/69
