+++ helion/language/atomic_ops.py
@@ -59,45 +59,6 @@ def _prepare_mem_args(
     return (target, index, *values, sem)
 
 
-def _codegen_common(
-    op: str, state: CodegenState, value_exprs: list[ast.AST]
-) -> ast.AST:
-    """Route any single-value atomic op through the atomic_indexing strategy."""
-    target = state.proxy_arg(0)
-    index = state.proxy_arg(1)
-    sem = expr_from_string(repr(state.proxy_arg(len(state.ast_args) - 1)))
-
-    assert isinstance(target, torch.Tensor)
-    assert isinstance(index, list)
-
-    host_function = HostFunction.current()
-    if target not in host_function.tensor_to_origin:
-        raise exc.AtomicOnDeviceTensor(op)
-
-    device_fn = state.device_function
-    fx_node = state.fx_node
-    epilogue_subtile_group_id = (
-        None if fx_node is None else fx_node.meta.get("epilogue_subtile_group_id")
-    )
-    if epilogue_subtile_group_id is None:
-        indexing_idx = device_fn.atomic_op_index
-        device_fn.atomic_op_index += 1
-    elif fx_node is not None and fx_node.meta.get(
-        "epilogue_subtile_primary_output", False
-    ):
-        indexing_idx = device_fn.atomic_op_index
-        device_fn.atomic_op_index += 1
-        device_fn.epilogue_subtile_atomic_indices[epilogue_subtile_group_id] = (
-            indexing_idx
-        )
-    else:
-        indexing_idx = device_fn.epilogue_subtile_atomic_indices[
-            epilogue_subtile_group_id
-        ]
-    strategy = device_fn.get_atomic_indexing_strategy(indexing_idx)
-    return strategy.codegen_atomic(op, state, target, index, value_exprs[0], sem)
-
-
 def _cute_pointer_expr(
     state: CodegenState,
     target: torch.Tensor,
@@ -1004,12 +965,6 @@ def _(
     return prev
 
 
-@_decorators.codegen(atomic_add, "triton")
-def _(state: CodegenState) -> ast.AST:
-    value_expr = state.ast_args[2]
-    return _codegen_common("atomic_add", state, _to_ast_values([value_expr]))
-
-
 @_decorators.codegen(atomic_add, "cute")
 def _(state: CodegenState) -> ast.AST:
     value_expr = state.ast_args[2]
@@ -1077,12 +1032,6 @@ def _(
     return _ref_atomic_binop(target, index, value, lambda old, val: val)
 
 
-@_decorators.codegen(atomic_xchg, "triton")
-def _(state: CodegenState) -> ast.AST:
-    value_expr = state.ast_args[2]
-    return _codegen_common("atomic_xchg", state, _to_ast_values([value_expr]))
-
-
 @_decorators.codegen(atomic_xchg, "cute")
 def _(state: CodegenState) -> ast.AST:
     value_expr = state.ast_args[2]
@@ -1147,12 +1096,6 @@ def _(
     return _ref_atomic_binop(target, index, value, torch.bitwise_and)
 
 
-@_decorators.codegen(atomic_and, "triton")
-def _(state: CodegenState) -> ast.AST:
-    value_expr = state.ast_args[2]
-    return _codegen_common("atomic_and", state, _to_ast_values([value_expr]))
-
-
 @_decorators.codegen(atomic_and, "cute")
 def _(state: CodegenState) -> ast.AST:
     value_expr = state.ast_args[2]
@@ -1214,12 +1157,6 @@ def _(
     return _ref_atomic_binop(target, index, value, torch.bitwise_or)
 
 
-@_decorators.codegen(atomic_or, "triton")
-def _(state: CodegenState) -> ast.AST:
-    value_expr = state.ast_args[2]
-    return _codegen_common("atomic_or", state, _to_ast_values([value_expr]))
-
-
 @_decorators.codegen(atomic_or, "cute")
 def _(state: CodegenState) -> ast.AST:
     value_expr = state.ast_args[2]
@@ -1279,12 +1216,6 @@ def _(
 ) -> torch.Tensor:
     _validate_sem(sem)
     return _ref_atomic_binop(target, index, value, torch.bitwise_xor)
-
-
-@_decorators.codegen(atomic_xor, "triton")
-def _(state: CodegenState) -> ast.AST:
-    value_expr = state.ast_args[2]
-    return _codegen_common("atomic_xor", state, _to_ast_values([value_expr]))
 
 
 @_decorators.codegen(atomic_xor, "cute")
@@ -1352,12 +1283,6 @@ def _(
     return _ref_atomic_binop(target, index, value, torch.maximum)
 
 
-@_decorators.codegen(atomic_max, "triton")
-def _(state: CodegenState) -> ast.AST:
-    value_expr = state.ast_args[2]
-    return _codegen_common("atomic_max", state, _to_ast_values([value_expr]))
-
-
 @_decorators.codegen(atomic_max, "cute")
 def _(state: CodegenState) -> ast.AST:
     value_expr = state.ast_args[2]
@@ -1418,12 +1343,6 @@ def _(
 ) -> torch.Tensor:
     _validate_sem(sem)
     return _ref_atomic_binop(target, index, value, torch.minimum)
-
-
-@_decorators.codegen(atomic_min, "triton")
-def _(state: CodegenState) -> ast.AST:
-    value_expr = state.ast_args[2]
-    return _codegen_common("atomic_min", state, _to_ast_values([value_expr]))
 
 
 @_decorators.codegen(atomic_min, "cute")
@@ -1535,35 +1454,6 @@ def _(
     return prev
 
 
-@_decorators.codegen(atomic_cas, "triton")
-def _(state: CodegenState) -> ast.AST:
-    exp_expr = state.ast_args[2]
-    val_expr = state.ast_args[3]
-    target = state.proxy_arg(0)
-    index = state.proxy_arg(1)
-    sem = expr_from_string(repr(state.proxy_arg(len(state.ast_args) - 1)))
-
-    assert isinstance(target, torch.Tensor)
-    assert isinstance(index, list)
-
-    # CAS always uses pointer (not a TMA reduction op, two values),
-    # but increment the counter to keep per-op atomic_indexing aligned.
-    device_fn = state.device_function
-    device_fn.atomic_op_index += 1
-
-    indices = SubscriptIndexing.create(state, target, index)
-    name = state.device_function.tensor_arg(target).name
-
-    exp_ast, val_ast = _to_ast_values([exp_expr, val_expr])
-    return expr_from_string(
-        f"tl.atomic_cas({name} + {{offset}}, {{exp}}, {{val}}, sem={{sem}})",
-        offset=indices.index_expr,
-        exp=exp_ast,
-        val=val_ast,
-        sem=sem,
-    )
-
-
 @_decorators.codegen(atomic_cas, "cute")
 def _(state: CodegenState) -> ast.AST:
     exp_expr = state.ast_args[2]
@@ -1611,3 +1501,4 @@ ATOMIC_OPS: frozenset[Callable[..., object]] = frozenset(
 # @_decorators.codegen(op, "<backend>") registrations run with the same eager
 # timing as when the bodies lived in this file -- no behavior change.
 import helion._compiler.pallas.atomic_ops  # noqa: E402, F401
+import helion._compiler.triton.atomic_ops  # noqa: E402, F401
