+++ helion/language/debug_ops.py
@@ -10,7 +10,6 @@ from .. import exc
 from . import _decorators
 
 if TYPE_CHECKING:
-    from .._compiler.inductor_lowering import CodegenState
     from .._compiler.type_info import TypeInfo
     from .._compiler.variable_origin import Origin
 
@@ -47,11 +46,14 @@ def _(*args: object, origin: Origin, **kwargs: object) -> TypeInfo:
     return LiteralType(origin, None)
 
 
-@_decorators.codegen(breakpoint, "triton")
-def _(state: CodegenState) -> None:
-    state.add_statement("breakpoint()")
-
-
 @_decorators.ref(breakpoint)
 def _() -> None:
     builtins.breakpoint()
+
+
+# ---------------------------------------------------------------------------
+# Backend-specific codegens for these ops live in per-backend modules under
+# helion/_compiler/<backend>/.  Import them here (at module import time) so the
+# @_decorators.codegen(op, "<backend>") registrations run with the same eager
+# timing as when the bodies lived in this file -- no behavior change.
+import helion._compiler.triton.debug_ops  # noqa: E402, F401
