+++ helion/language/quantized_ops.py
@@ -1,16 +1,16 @@
 from __future__ import annotations
 
-import ast
 from typing import TYPE_CHECKING
 
 import torch
 
 from .. import exc
-from .._compiler.ast_extension import create
 from .._compiler.ast_extension import expr_from_string
 from . import _decorators
 
 if TYPE_CHECKING:
+    import ast
+
     from .._compiler.inductor_lowering import CodegenState
 
 __all__ = ["float4_e2m1fn_x2_to_float32"]
@@ -50,43 +50,9 @@ def _(state: CodegenState) -> list[ast.AST]:
     ]
 
 
-# Triton uses inline asm for the same fp4x2 unpack instruction as CuTe.
-_FP4X2_TO_F32_ASM = """
-{
-    .reg .b8 b0;
-    .reg .b16 lo, hi;
-    .reg .b32 h2;
-    mov.b16 {b0, _}, $2;
-    cvt.rn.f16x2.e2m1x2 h2, b0;
-    mov.b32 {lo, hi}, h2;
-    cvt.f32.f16 $0, lo;
-    cvt.f32.f16 $1, hi;
-}
-"""
-
-
-@_decorators.codegen(float4_e2m1fn_x2_to_float32, "triton")
-def _(state: CodegenState) -> list[ast.AST]:
-    value = state.codegen.lift(
-        expr_from_string("{value}.to(tl.int16)", value=state.ast_arg(0)),
-        dce=True,
-        prefix="fp4_packed",
-    )
-    call = create(
-        ast.Call,
-        func=expr_from_string("tl.inline_asm_elementwise"),
-        args=[
-            create(ast.Constant, value=_FP4X2_TO_F32_ASM),
-            create(ast.Constant, value="=f,=f,h"),
-            create(ast.List, elts=[value], ctx=ast.Load()),
-            expr_from_string("(tl.float32, tl.float32)"),
-            create(ast.Constant, value=True),  # is_pure
-            create(ast.Constant, value=1),  # pack
-        ],
-        keywords=[],
-    )
-    result = state.codegen.lift(call, dce=True, prefix="fp4_pair")
-    return [
-        expr_from_string(f"{result.id}[0]"),
-        expr_from_string(f"{result.id}[1]"),
-    ]
+# ---------------------------------------------------------------------------
+# Backend-specific codegens for these ops live in per-backend modules under
+# helion/_compiler/<backend>/.  Import them here (at module import time) so the
+# @_decorators.codegen(op, "<backend>") registrations run with the same eager
+# timing as when the bodies lived in this file -- no behavior change.
+import helion._compiler.triton.quantized_ops  # noqa: E402, F401
