+++ helion/language/scan_ops.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import ast
 import operator
 from typing import TYPE_CHECKING
 from typing import cast
@@ -14,6 +13,8 @@ from .. import exc
 from . import _decorators
 
 if TYPE_CHECKING:
+    import ast
+
     from .._compiler.device_ir import HelperFunctionGraphInfo
     from .._compiler.helper_function import CombineFunction
     from .._compiler.helper_function import CombineFunctionBasic
@@ -354,26 +355,6 @@ def _(
         return tuple(torch.empty_like(t) for t in input_tensor)
     assert isinstance(input_tensor, torch.Tensor), input_tensor
     return torch.empty_like(input_tensor)
-
-
-@_decorators.codegen(_associative_scan, "triton")
-def _(state: CodegenState) -> ast.AST | list[ast.AST]:
-    """Generate code for associative scan with combine function."""
-
-    combine_graph_id = state.proxy_arg(0)
-    dim = state.proxy_arg(2)
-    reverse = state.proxy_arg(3)
-    is_tuple_input = state.proxy_arg(4)
-
-    input_tensor = _get_input_tensor_ast(state, bool(is_tuple_input))
-    helper_func_name = _register_helper_function(state, cast("int", combine_graph_id))
-    scan_expr = _create_scan_expression(
-        input_tensor, cast("int", dim), helper_func_name, bool(reverse)
-    )
-
-    if is_tuple_input:
-        return _create_tuple_result_expressions(state, scan_expr)
-    return scan_expr
 
 
 @_decorators.codegen(_associative_scan, "cute")
@@ -924,23 +905,6 @@ def _cute_codegen_serial_scan(
     return [expr_from_string(acc_var) for acc_var in acc_vars]
 
 
-def _get_input_tensor_ast(state: CodegenState, is_tuple_input: bool) -> ast.AST:
-    """Get the input tensor AST, handling tuple inputs specially."""
-    if not is_tuple_input:
-        return state.ast_arg(1)
-
-    raw_input = state.ast_args[1]
-    if isinstance(raw_input, tuple):
-        from .._compiler.ast_extension import create
-
-        tuple_elts = [
-            elt if isinstance(elt, ast.AST) else ast.Constant(value=elt)
-            for elt in raw_input
-        ]
-        return create(ast.Tuple, elts=tuple_elts, ctx=ast.Load())
-    return state.ast_arg(1)
-
-
 def _scan_combine_operator(helper_graph_info: HelperFunctionGraphInfo) -> str:
     import operator as operator_mod
 
@@ -1028,49 +992,9 @@ def _cute_sorted_value_lines(
     ]
 
 
-def _register_helper_function(state: CodegenState, combine_graph_id: int) -> str:
-    """Register the helper function and return its final name."""
-    from .._compiler.device_ir import HelperFunctionGraphInfo
-
-    helper_graph_info = state.get_graph(combine_graph_id)
-    assert isinstance(helper_graph_info, HelperFunctionGraphInfo)
-    state.codegen.device_function.register_helper_function(helper_graph_info)
-    # Get the final name from the helper manager (which uses the namespace)
-    return state.codegen.device_function.helper_manager.get_final_name(
-        helper_graph_info
-    )
-
-
-def _create_scan_expression(
-    input_tensor: ast.AST, dim: int, helper_func_name: str, reverse: bool
-) -> ast.AST:
-    """Create the tl.associative_scan expression."""
-    from .._compiler.ast_extension import expr_from_string
-
-    template = (
-        f"tl.associative_scan({{input_tensor}}, {{dim_value}}, {helper_func_name}, reverse=True)"
-        if reverse
-        else f"tl.associative_scan({{input_tensor}}, {{dim_value}}, {helper_func_name})"
-    )
-    return expr_from_string(
-        template,
-        input_tensor=input_tensor,
-        dim_value=ast.Constant(value=dim),
-    )
-
-
-def _create_tuple_result_expressions(
-    state: CodegenState, scan_expr: ast.AST
-) -> list[ast.AST]:
-    """Create getitem expressions for tuple results."""
-    from .._compiler.ast_extension import expr_from_string
-
-    raw_input = state.ast_args[1]
-    num_elements = len(raw_input) if isinstance(raw_input, tuple) else 2
-
-    return [
-        expr_from_string(
-            "{scan_result}[{index}]", scan_result=scan_expr, index=ast.Constant(value=i)
-        )
-        for i in range(num_elements)
-    ]
+# ---------------------------------------------------------------------------
+# Backend-specific codegens for these ops live in per-backend modules under
+# helion/_compiler/<backend>/.  Import them here (at module import time) so the
+# @_decorators.codegen(op, "<backend>") registrations run with the same eager
+# timing as when the bodies lived in this file -- no behavior change.
+import helion._compiler.triton.scan_ops  # noqa: E402, F401
